pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentiment_analysis"
version = "1.0.0"
dependencies = [
    "flwr[simulation]>=1.12.0",
    "flwr-datasets[vision]>=0.3.0",
    "torch==2.2.1",
    "transformers",
    "matplotlib",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "flwrlabs"

[tool.flwr.app.components]
serverapp = "sentiment_analysis.server_app:app"
clientapp = "sentiment_analysis.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 2
fraction-fit = 0.25
fraction-evaluate = 0.25
server-device = "cpu"

[tool.flwr.federations]
default = "local-sim-gpu"

[tool.flwr.federations.local-sim]
options.num-supernodes = 4
options.backend.client-resources.num-cpus = 4
options.backend.client-resources.num-gpus = 0.0

[tool.flwr.federations.local-sim-gpu]
options.num-supernodes = 4
options.backend.client-resources.num-cpus = 4
options.backend.client-resources.num-gpus = 0.5
# options.backend.init_args.num_cpus = 1
# options.backend.init_args.num_gpus = 1